% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSTestInit.R
\name{PSTestInit}
\alias{PSTestInit}
\title{Initialise the data for performing the Monte Carlo test for preferential sampling

\code{PSTestInit} returns the data in the neccessary format for use with the function
  \code{\link{PSTestRun}}.}
\usage{
PSTestInit(
  type,
  discrete,
  positions = NULL,
  times = NULL,
  poly = NULL,
  discrete_locations = NULL,
  areal_polygons = NULL,
  areal_poly_observations = NULL,
  n_prediction = 10000
)
}
\arguments{
\item{type}{is a character string taking value 'spatial' or 'spacetime'.}

\item{discrete}{is a logical value stating if the data are discrete spatial
(i.e. areal data), or continous spatial (i.e. point-referenced data).}

\item{positions}{give the coordinates of the observations in space and time.
This argument only needs specifying when discrete==F. When discrete=F,
this is either an object of class ppp or else it takes a data.frame with
named values x and y. These give the coordinates of the observed points.}

\item{times}{give the discrete times of the observed points. Only required
in the spacetime setting (i.e. when type=='spacetime').}

\item{poly}{this takes a polygon-style object defining the boundaries of
the study region. This can be of class owin (from the spatstat package),
SpatialPolygons (from the sp package) or sf (from the sf package).
This argument is only required when when discrete==F.}

\item{discrete_locations}{is an optional argument when discrete==T. It gives
the locations of the population of areal units. This is useful to specify
when the distances with respect to areal unit centroids is not desirable.}

\item{areal_polygons}{Specifies the polygons defining the population of areal
units. This argument only needs specifying when discrete==T.
This takes an object of class owin (from the spatstat package),
SpatialPolygons (from the sp package) or sf (from the sf package).}

\item{areal_poly_observations}{this is a vector of indices denoting the
polygons with complete data. This argument only needs specifying when
discrete==T.}

\item{n_prediction}{is an integer specifying the number of prediction
locations to generate within poly. This argument only needs specifying when
discrete==T.}
}
\value{
A named list for use with \code{\link{PSTestRun}}.
}
\description{
Initialise the data for performing the Monte Carlo test for preferential sampling

\code{PSTestInit} returns the data in the neccessary format for use with the function
  \code{\link{PSTestRun}}.
}
\section{Examples}{

  For detailed examples, see the vignette, or visit
  \url{https://github.com/joenomiddlename/PStestR} for more details.
}

\seealso{
\code{\link{PSTestRun}}
}
