% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsum_functions.R
\name{rayleigh.opt}
\alias{rayleigh.opt}
\title{Collection of helper functions for main simulation}
\usage{
rayleigh.opt(u = NULL, v = NULL, par = NULL)
}
\description{
Collection of helper functions for main simulation
}
