% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsum_altsim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Sets up necessary parameters for simulation}
\usage{
sim_setup(n = 100, p = 1000, model = "normal", sigma = 1,
  nsim = 200, alpha = 0.05, seed = 2019, rho = 0.9, betasp = 1,
  rs = c(10, 20, 50))
}
\description{
Sets up necessary parameters for simulation
}
