% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsum_altsim.R
\name{pstest}
\alias{pstest}
\title{Performs the power analysis for the projected score test}
\usage{
pstest(nsim = 500, seed = 2019, mbeta = 0, kperc = 40, n = 100,
  p = 1000, model = "normal", sigma = 1, alpha = 0.05,
  betasp = TRUE, rs = c(10, 20, 50), mc.cores = 1, Gprime, GQs, GQs2,
  R1nams, R2nams, nams, H1, A, G, linkatlambda, simresults, powresults)
}
\arguments{
\item{nsim}{number of simulations to run to calculate power}

\item{seed}{set a seed for the power calculation, defaults to 2019}

\item{mbeta}{mean coefficient for nonzero effects, defaults to 0}

\item{kperc}{percentage of independent variables with nonzero signal, defaults to 40}

\item{n}{defaults to 100}

\item{p}{defaults to 1000}

\item{model}{can be specified as 'normal' (default) for linear regression, otherwise does logistic regression}

\item{sigma}{defaults to 1}

\item{alpha}{significance level, defaults to 0.05}

\item{betasp}{indicator of presence of spatial information, defaults to TRUE}

\item{rs}{investigator-specified set of "contrasts" of G, defaults to c(10, 20, 50)}

\item{mc.cores}{number of cores to run simulation on}

\item{Gprime}{parameter from sim_setup(), defaults to NULL}

\item{GQs}{parameter from sim_setup(), defaults to NULL}

\item{GQs2}{parameter from sim_setup(), defaults to NULL}

\item{R1nams}{parameter from sim_setup(), defaults to NULL}

\item{R2nams}{parameter from sim_setup(), defaults to NULL}

\item{nams}{parameter from sim_setup(), defaults to NULL}

\item{H1}{parameter from sim_setup(), defaults to NULL}

\item{A}{parameter from sim_setup(), defaults to NULL}

\item{G}{parameter from sim_setup(), defaults to NULL}

\item{linkatlambda}{parameter from sim_setup(), defaults to NULL}

\item{simresults}{empty dataframe from sim_setup(), defaults to NULL}

\item{powresults}{empty dataframe from sim_setup(), defaults to NULL}
}
\value{
A list of the entire matrix of simulation results, the power results
}
\description{
This is an independent function which calls setup from maxsum_altsim_setup.R,
which sets up all of the variables needed for one value of kperc, percentage of independent variables with nonzero signal
and one value of mbeta, mean coefficient for non zero effects uniform around that value
That function loops through the variables k and beta.
}
\keyword{projected}
\keyword{score}
\keyword{test}
