% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsum_functions.R
\name{rayleigh.opt}
\alias{rayleigh.opt}
\title{Rayleigh Quotient}
\usage{
rayleigh.opt(u, v, par)
}
\arguments{
\item{u}{some matrix}

\item{v}{some matrix}

\item{par}{nonzero vector x}
}
\description{
function to estimate alpha vector
This optimizes a Rayleigh quotient which is the solution
to a generalized eigenvalue problem. alpha is the first
eigenvector.
}
\details{
Collection of helper functions for main simulation
}
