% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsum_altsim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Sets up necessary parameters for power calculation simulation}
\usage{
sim_setup(nsim = 500, seed = 2019, n = 100, p = 1000,
  model = "normal", sigma = 1, rho = 0.9, rs = c(10, 20, 50))
}
\arguments{
\item{nsim}{defaults to 500}

\item{seed}{set a seed for the generation of matrices needed for the various tests, defaults to 2019}

\item{n}{defaults to 100}

\item{p}{defaults to 1000}

\item{model}{can be specified as 'normal' (default) for linear regression, otherwise does logistic regression}

\item{sigma}{defaults to 1}

\item{rho}{spatial correlation in G parameter, AR1 structure, defaults to 0.9}

\item{rs}{investigator-specified set of "contrasts" of G, defaults to c(10, 20, 50)}
}
\value{
A list including spatial correlation parameters, empty dataframes for simresults
 and powresults, parameters for distribution of SKAT statistic
}
\description{
Support function for pstest(), which runs the power calculations that are wrapped in pst_sim(). Sets
up objects needed to do power calculation.
}
